refactor in prep for rewrite
