+++ mtriage
@@ -204,78 +204,8 @@ def build(args):
     os.remove(BUILD_PIPFILE)
 
 
-def develop(args):
-    TAG_NAME = "dev-gpu" if args.gpu else "dev"
-    # --runtime only exists on nvidia docker, so we pass a bubblegum flag when not available
-    # so that the call arguments are well formed.
-    sp.call(
-        [
-            "docker",
-            "run",
-            "-it",
-            "--name",
-            CONT_NAME,
-            "--runtime=nvidia" if args.gpu else "--ipc=host",
-            "--env",
-            "BASE_DIR=/mtriage",
-            "--env-file={}".format(ENV_FILE),
-            "--rm",
-            "--privileged",
-            "-v",
-            "{}:/mtriage".format(DIR_PATH),
-            "-v",
-            "{}/.config/gcloud:/root/.config/gcloud".format(HOME_PATH),
-            "{}:{}".format(NAME, TAG_NAME),
-        ]
-    )
-
-
-def clean(args):
-    sp.call(["docker", "rmi", NAME])
-
-
-def __run_lib_tests():
-    returncode = sp.call(
-        [
-            "docker",
-            "run",
-            "--env",
-            "BASE_DIR=/mtriage",
-            "--env-file={}".format(ENV_FILE),
-            "--rm",
-            "-v",
-            "{}:/mtriage".format(DIR_PATH),
-            "--workdir",
-            "/mtriage/src",
-            "{}:dev".format(NAME),
-            "python",
-            "-m",
-            "pytest",
-            "test",
-        ]
-    )
-    if returncode is 1:
-        exit(returncode)
-
-
-def __run_runpy_tests():
-    # NOTE: runpy tests are not run in a docker container, as they operate on the local machine-- so this test is run
-    # using the LOCAL python (could be 2 or 3).
-    returncode = sp.call(["python", "-m", "pytest", "test/"])
-    if returncode is 1:
-        exit(returncode)
-
-
-def test(args):
-    print("Creating container to run tests...")
-    print("----------------------------------")
-    __run_lib_tests()
-    __run_runpy_tests()
-    print("----------------------------------")
-    print("All tests for mtriage done.")
-
-
 def verify_viewer_args(input_dir, all_viewers_dir, viewer_name):
+    # TODO: adjust and use before running docker commmands
     if input_dir == None:
         raise InvalidArgumentsError("No input directory supplied for viewer plugin.")
 
@@ -301,47 +231,6 @@ def verify_viewer_args(input_dir, all_viewers_dir, viewer_name):
         )
 
     return viewer_config_path
-
-
-def create_symlinks(input_elements_dir, server_elements_dir):
-    # clean if exists
-    if not os.path.exists(input_elements_dir):
-        raise OSError("Server elements dir does not exist")
-
-    # make it empty
-    shutil.rmtree(server_elements_dir)
-    os.makedirs(server_elements_dir)
-
-    el_ids = [
-        f
-        for f in os.listdir(input_elements_dir)
-        if os.path.isdir("{}/{}".format(input_elements_dir, f))
-    ]
-
-    # create symlinks for input elements
-    for el_id in el_ids:
-        folder_in = "{}/{}".format(input_elements_dir, el_id)
-        for element_path in os.listdir(folder_in):
-            folder_out = "{}/{}".format(server_elements_dir, el_id)
-            folder_out_media = "{}/media".format(folder_out)
-            if not os.path.exists(folder_out):
-                os.makedirs(folder_out_media)
-            os.symlink(
-                "{}/{}".format(folder_in, element_path),
-                "{}/{}".format(folder_out_media, element_path),
-            )
-
-
-def get_viewer_etype(viewer_config_path):
-    with open(viewer_config_path, "r") as f:
-        viewer_config = json.load(f)
-        viewer_etype = viewer_config["etype"]
-        return viewer_etype
-
-
-def create_server_config(configPath, configDict):
-    with open(configPath, "w") as config:
-        json.dump(configDict, config)
 
 
 def __build_viewer_server():
@@ -388,29 +277,13 @@ def __build_viewer_client(viewer_name):
     print("----------------------------------")
 
 
-def __setup_symlinks(input_elements_dir, viewer_name):
-    all_viewers_dir = "src/lib/viewers"
-    viewer_config_path = verify_viewer_args(
-        input_elements_dir, all_viewers_dir, viewer_name
-    )
-    server_elements_dir = "src/server/elements"
-    server_config_path = "src/server/config.json"
-    viewer_etype = get_viewer_etype(viewer_config_path)
-
-    server_config = {"port": 8080, "etype": viewer_etype}
-    viewer_img_name = "{}:{}".format(VIEWER_NAME, viewer_name)
-
-    create_server_config(server_config_path, server_config)
-    create_symlinks(input_elements_dir, server_elements_dir)
-
-
 def view(args):
-    """ Must be invoked with an input folder and viewer-plugin e.g.:
+    """ Must be invoked with an input working folder and viewer-plugin e.g.:
 
-            ./mtriage view -i <derived_folder> -v <viewer_plugin>
+            ./mtriage view -i <working_folder> -v <viewer_plugin>
 
-        Server requires that input folder contain directories corresponsing to elements (where directory names
-        are element ids.)
+        The input folder should be an mtriage working directory. Mserver will index and present both selected and
+        analysed element folders.
 
         Viewers must be yarn buildable apps living in the src/lib/viewers folder. Their name is taken to be the name
         of their outer directory. Viewers must specify which etype they accept in their config.yml file.
@@ -453,7 +326,6 @@ def view(args):
     """
     input_elements_dir = args.input
     viewer_name = args.viewer
-    __setup_symlinks(input_elements_dir, viewer_name)
 
     is_local = True
     server_img_name = "{}:dev".format(SERVER_NAME)
@@ -545,6 +417,77 @@ def dev_viewerserver(args):
 
     signal.signal(signal.SIGINT, signal_handler)
     signal.pause()
+
+
+def develop(args):
+    TAG_NAME = "dev-gpu" if args.gpu else "dev"
+    # --runtime only exists on nvidia docker, so we pass a bubblegum flag when not available
+    # so that the call arguments are well formed.
+    sp.call(
+        [
+            "docker",
+            "run",
+            "-it",
+            "--name",
+            CONT_NAME,
+            "--runtime=nvidia" if args.gpu else "--ipc=host",
+            "--env",
+            "BASE_DIR=/mtriage",
+            "--env-file={}".format(ENV_FILE),
+            "--rm",
+            "--privileged",
+            "-v",
+            "{}:/mtriage".format(DIR_PATH),
+            "-v",
+            "{}/.config/gcloud:/root/.config/gcloud".format(HOME_PATH),
+            "{}:{}".format(NAME, TAG_NAME),
+        ]
+    )
+
+
+def clean(args):
+    sp.call(["docker", "rmi", NAME])
+
+
+def __run_lib_tests():
+    returncode = sp.call(
+        [
+            "docker",
+            "run",
+            "--env",
+            "BASE_DIR=/mtriage",
+            "--env-file={}".format(ENV_FILE),
+            "--rm",
+            "-v",
+            "{}:/mtriage".format(DIR_PATH),
+            "--workdir",
+            "/mtriage/src",
+            "{}:dev".format(NAME),
+            "python",
+            "-m",
+            "pytest",
+            "test",
+        ]
+    )
+    if returncode is 1:
+        exit(returncode)
+
+
+def __run_runpy_tests():
+    # NOTE: runpy tests are not run in a docker container, as they operate on the local machine-- so this test is run
+    # using the LOCAL python (could be 2 or 3).
+    returncode = sp.call(["python", "-m", "pytest", "test/"])
+    if returncode is 1:
+        exit(returncode)
+
+
+def test(args):
+    print("Creating container to run tests...")
+    print("----------------------------------")
+    __run_lib_tests()
+    __run_runpy_tests()
+    print("----------------------------------")
+    print("All tests for mtriage done.")
 
 
 if __name__ == "__main__":
